+++ .vscode/launch.json
@@ -5,6 +5,21 @@
     "version": "0.2.0",
     "configurations": [
         {
+            "name": "Debug Unit Test",
+            "type": "debugpy",
+            "request": "launch",
+            "justMyCode": false,
+            "program": "${file}",
+            "purpose": [
+                "debug-test"
+            ],
+            "console": "integratedTerminal",
+            "env": {
+                // Enable break on exception when debugging tests (see: tests/conftest.py)
+                "PYTEST_RAISE": "1",
+            },
+        },
+        {
             "name": "panda",
             "type": "debugpy",
             "request": "launch",
@@ -15,6 +30,6 @@
                 "run",
                 "${workspaceFolder:fastcs-PandABlocks}/src/fastcs_pandablocks/panda/controller.yaml"
             ]
-        }
+        },
     ]
 }
